backend/requirements.txt
alembic==1.0.5
aniso8601==3.0.2
asn1crypto==0.24.0
cffi==1.11.5
Click==7.0
cryptography==2.4.2
email-validator==1.1.1
Flask==1.0.2
Flask-Cors==3.0.7
Flask-GraphQL==2.0.0
Flask-Migrate==2.3.1
Flask-MySQLdb==0.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
graphene==2.1.3
graphene-file-upload==1.2.2
graphene-sqlalchemy==2.1.0
graphql-core==2.1
graphql-relay==0.4.5
graphql-server-core==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
mysqlclient==2.0.1
promise==2.2.1
PyJWT==1.7.1
python-dateutil==2.7.5
PyYAML==5.3.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
boto3==1.14.47
botocore==1.17.47
jmespath==0.10.0
Werkzeug==0.14
SQLAlchemy==1.3.16
python-editor==1.0.4
pycparser==2.20
PyMySQL
Pillow==7.2.0
